refactoring startup and cli
